refactor useInsightsSQLEditorOnMountCallback folder structure
