+++ ui/apps/dashboard/src/components/Insights/InsightsSQLEditor/hooks/useLatestCallback.ts
